apitests/requirements.txt
pytest
requests
testcontainers
testcontainers[opensearch]
testcontainers[postgresql]
testcontainers[nats]
openai
pytest_httpserver
websocket-client
